vagrant: Support sshfs, move sync to admin1 only, change paths
sshfs is a lot nicer than rsync, so default to it if available.  Also,
we only need to do the sync to admin1, so move the definitions there.

At the same time, I changed the default paths because on Atomic Host
it's not permitted by default to create `/vagrant`.  We use
`/home/vagrant/sync` instead.